# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_chizze_backend_internal_models.CreateReviewRequest:
    properties:
      delivery_rating:
        maximum: 5
        minimum: 1
        type: integer
      food_rating:
        maximum: 5
        minimum: 1
        type: integer
      review_text:
        type: string
      tags:
        items:
          type: string
        type: array
    required:
    - food_rating
    type: object
...
  github_com_chizze_backend_internal_models.ReplyReviewRequest:
    properties:
      reply:
        type: string
    required:
    - reply
    type: object
...
host: api.devdeepak.me
info:
  contact:
    email: support@chizze.app
    name: Chizze Support
    url: https://chizze.app/support
  description: Backend API for Chizze — a food delivery platform with customer, restaurant
    partner, and delivery partner roles.
  license:
    name: Proprietary
    url: https://chizze.app/license
  termsOfService: https://chizze.app/terms
  title: Chizze Food Delivery API
  version: "1.0"
paths:
...
  /api/v1/orders/{id}/review:
    post:
      consumes:
      - application/json
      description: Submits a review for a delivered order. Validates ownership, delivery
        status, and duplicate checks.
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      - description: Review data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_chizze_backend_internal_models.CreateReviewRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a review
      tags:
      - Reviews
...
  /api/v1/partner/reviews/{id}/reply:
    post:
      consumes:
      - application/json
      description: Allows a restaurant owner to reply to a review with ownership verification
      parameters:
      - description: Review ID
        in: path
        name: id
        required: true
        type: string
      - description: Reply text
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_chizze_backend_internal_models.ReplyReviewRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Reply to a review
      tags:
      - Reviews
...
securityDefinitions:
  BearerAuth:
    description: Enter "Bearer {token}"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"